Uploaded my Year 13 project
In Year 13 we had to make a program based on our chosen concept (AS 3.3 AND 3.46). My concept was to make a student timetable. Being ambitious, I thought I could do it but I only had half a year to make my program and we were all self-teaching ourselves Python. I did not have time to complete my concept due to also being preoccupied with my other classes. Hence, it is unfinished, but it still functions.